pyproject.toml
[project]
name = "keyvalstore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google",
    "grpc-tools",
    "grpcio>=1.76.0",
]
